缓存预热与动态 TTL

+++ backend/alex-seckill/alex-seckill-common/src/main/java/com/nextalex/seckill/common/domain/dataobject/SeckillGoodsDO.java
@@ -28,7 +28,7 @@ public class SeckillGoodsDO {
 
     private BigDecimal seckillTotal;
 
-    private Long seckillStock;
+    private Integer seckillStock;
 
     private Integer seckillLimit;
