+++ backend/alex-seckill/alex-seckill-goods/src/main/java/com/nextalex/seckill/goods/model/vo/FindSeckillGoodsDetailRspVO.java
@@ -63,7 +63,7 @@ public class FindSeckillGoodsDetailRspVO {
     /**
      * 秒杀商品剩余库存
      */
-    private Long seckillStock;
+    private Integer seckillStock;
 
     /**
      * 活动状态：0=未开始，1=进行中，2=已结束
